==> 00-0_find_knockouts.ipynb <==
#### _Matin's Lab Rotation 3 - BCFO_ | [NEXT](00-1_build_strain.ipynb)
# 01. Find Genes to Knockout
* In this first part, we want to find which genes need to be knocked out from the model
* List of knockout genes provided by Yixin:

## Xylitol strain
### Knock outs
* focA, https://www.uniprot.org/uniprot/P0AC23
* pflB, https://www.uniprot.org/uniprot/P09373
* ldhA, https://www.uniprot.org/uniprot/P52643
* adhE, https://www.uniprot.org/uniprot/P0A9Q7
* xylAB, https://www.uniprot.org/uniprot/D1MDQ9 or https://www.uniprot.org/uniprot/P09099?
* frdA, https://www.uniprot.org/uniprot/P00363
* cyoB, https://www.uniprot.org/uniprot/P0ABI8
* appB, https://www.uniprot.org/uniprot/P26458

### Missing reaction(s)
* xylose + NADPH -> xylitol + NADP+ https://www.brenda-enzymes.org/enzyme.php?ecno=1.1.1.307
add this reactaion https://www.uniprot.org/uniprot/Q8X195

## Isobutyric acid strain
### Knock outs
* focA, https://www.uniprot.org/uniprot/P0AC23
* pflB, https://www.uniprot.org/uniprot/P09373
* aceEF, https://www.uniprot.org/uniprot/P0ACL9
* poxB, https://www.uniprot.org/uniprot/P07003
* tdcE, https://www.uniprot.org/uniprot/P42632
* pflDC, https://www.uniprot.org/uniprot/P32675 
* deoC, https://www.uniprot.org/uniprot/P0A6L0
* ydbK, https://www.uniprot.org/uniprot/P52647
* yqhD, https://www.uniprot.org/uniprot/Q46856
* araB, https://www.uniprot.org/uniprot/P08204
* araA, https://www.uniprot.org/uniprot/P08202
* xylAB, https://www.uniprot.org/uniprot/D1MDQ9


### Missing reactions(s): 
* isobutyraldehyde to isobutyric acid (forgot the enzyme/gene name)
Catalyzed by padA and has the following reaction (replace substrate and product):
* phenylacetaldehyde + NAD+ + H2O → phenylacetate + NADH + 2 H+
* Originally identified to oxidize isobutyraldehyde in this paper:
https://findit.dtu.dk/en/catalog/2482526722
https://www.uniprot.org/uniprot/P80668
http://bigg.ucsd.edu/models/iML1515/reactions/ALDD19xr
* Also add this gene:
kivD - https://www.uniprot.org/uniprot/D2BR82
https://biocyc.org/META/NEW-IMAGE?object=RXN-7643

# Write the knockout list in a table
Find the table in:
* [data_ko_xylitol_strain.txt](data/data_ko_xylitol_strain.txt)
* [data_ko_isobutyric-acid_strain](data/data_ko_isobutyric-acid_strain.txt)

# Load Library

from bioservices import UniProt
import io
import pandas as pd
import numpy as np

from cameo import models
from cameo import load_model

service = UniProt()

# get value from uniprot

def get_uniprot(path, strain_name):
    df_input = pd.read_csv(path, header = None,
                     skiprows = 0,
                     sep = ',',
                     names = [
                         'gene',
                         'uniprot',
                         'strain'])
    df_input['strain'] = strain_name
    df = pd.DataFrame(columns=["Entry", "Entry name", "Gene names  (primary )", "Gene names  (synonym )", "Protein names","Gene ontology (molecular function)","Catalytic activity"])
    for i in df_input.uniprot:
        uniprot_id = i.split('/')[-1]
        # Make a query string
        query = "accession:"+uniprot_id
        # Define a list of columns we want to retrive
        columnlist = "id,entry name,genes(PREFERRED),genes(ALTERNATIVE),protein names,go(molecular function),comment(CATALYTIC ACTIVITY)"
        # Run the remote search
        result = service.search(query, frmt="tab", columns=columnlist)
        df_add = (pd.read_table(io.StringIO(result)))
        df = pd.concat([df, df_add])
    df = df.reset_index(drop=True)
    df_input = pd.merge(df_input, df, left_index=True, right_index=True)
    return df_input

def check_reaction(df):
    reaction_data = []
    for counter, i in enumerate(df['Catalytic activity']):
        try:
            reaction = i.split(';')[0].replace('CATALYTIC ACTIVITY: Reaction=','')
            print(df.gene[counter]+':', reaction)
            reaction_data.append(reaction)
        except:
            print(df.gene[counter]+':', 'reaction not available')
            reaction_data.append(np.nan)
    df = df.assign(Reaction = reaction_data)
    return

# Find knockouts for xylitol strain

path = 'data/data_ko_xylitol_strain.txt'
strain_name = 'xylytol_mutant'
df_xylitol = get_uniprot(path, strain_name)
df_xylitol

check_reaction(df_xylitol)

# Find knockouts for isobutyric acid strain

path = 'data/data_ko_isobutyric-acid_strain.txt'
strain_name = 'isobutyric_mutant'
df_isobutyric = get_uniprot(path, strain_name)
df_isobutyric

check_reaction(df_isobutyric)

# Check if genes are in the model
## Which model should we use?
Here we get the model from BIGG and use E coli with the most reactions available

bigg_list = models.index_models_bigg()

# Find GEMS that starts with "Eschericia coli" and ends with "MG1655"
e_coli_mg1655 = bigg_list[bigg_list.organism.str.contains("^Escherichia coli.*MG1655$")]
e_coli_mg1655

model_id = e_coli_mg1655.iloc[e_coli_mg1655.reaction_count.argmax()].bigg_id
model_id

## Load Model

model = load_model(model_id)
model

# get all available genes and their index
gene_list = []
for counter, i in enumerate(model.genes):
    try:
        data = [str(i), counter, i.annotation['uniprot'][0]]
        gene_list.append(data)
    except:
        data = [str(i), counter, '']
        gene_list.append(data)
df_gene = pd.DataFrame(gene_list, columns=['id_bigg',model_id+'_idx','Entry'])
df_gene

## Find if the genes we want to knock out is in the model
### For Xylitol producing strain

df_xylitol = df_xylitol.join(df_gene.set_index('Entry'), on='Entry')
df_xylitol

#export the result
df_xylitol.to_csv('data/df_xylitol_knockouts.csv')

### For isobutyric acid producing strain

df_isobutyric = df_isobutyric.join(df_gene.set_index('Entry'), on='Entry')
df_isobutyric

#export the result
df_isobutyric.to_csv('data/df_isobutyric_knockouts.csv')

# [CLICK FOR NEXT STEP](00-1_build_strain.ipynb)
