==> 00-1_build_strain.ipynb <==
#### _Matin's Lab Rotation 3 - BCFO_ | [NEXT](01-0_simulation_xylitol_strain.ipynb) | [PREVIOUS](00-0_find_knockouts.ipynb)
# 02. Build KOs & Introduce Heterologous Pathway in the Model
List of the iML1515 BIGG Genes ID to be knocked out can be find in this table:
* [Xylitol strain](data/df_xylitol_knockouts.csv)
* [Isobutyric strain](data/df_isobutyric_knockouts.csv)

## Load Library

from cameo import models
from cameo import load_model
import pandas as pd
import numpy as np
from cobra import Model, Reaction, Metabolite
import cobra

def evaluate_knockout(model, mutant, ko_list):
    bound_knockout = []
    for i in ko_list:
        i_gene = str(model.genes[i])
        i_uniprot = str(model.genes[i].annotation['uniprot'][0])
        i_n_reactions = len(model.genes[i].reactions)
        i_reactions = []
        i_bounds_model = []
        i_bounds_mutant = []
        for x in model.genes[i].reactions:
            x_reaction = [str(x.id), str(x.reaction)]
            x_bounds = x.bounds
            i_reactions.append(x_reaction)
            i_bounds_model.append(x_bounds)
        for y in mutant.genes[i].reactions:
            y_bounds = y.bounds
            i_bounds_mutant.append(y_bounds)
        data = [i_gene, i_uniprot, i_n_reactions, i_reactions, i_bounds_model, i_bounds_mutant]
        bound_knockout.append(data)

    for i in bound_knockout:
        print('gene: '+i[0], '| uniprot: '+i[1])
        print('number of reactions involved: '+str(i[2]))
        #for count, x in enumerate(i[3]):
        #    print('* '+x[0]+' :', x[1])
        #    print('bounds normal: '+str(i[4][count]))
        #    print('bounds mutant: '+str(i[5][count]))
        print('------------------------------------------------------------')
    
    gene_reaction_pair = []
    for i in bound_knockout:
        gene_id = i[0]
        uniprot_id = i[1]
        for count, x in enumerate(i[3]):
            reaction_id = x[0]
            reaction = x[1]
            b_normal = i[4][count]
            b_mutant = i[5][count]
            if b_normal == b_mutant:
                ko_result = False
            else:
                ko_result = True
            pair = [uniprot_id, gene_id, reaction_id, reaction, b_normal, b_mutant, ko_result]
            gene_reaction_pair.append(pair)
    df = pd.DataFrame(data = gene_reaction_pair ,columns=["Entry", "id_BIGG", "reaction_id", "reaction", "normal_bound", "mutant_bound", "ko_result"])
    return df

## Load Model

model = load_model('iML1515')
model.solver = 'gurobi'
model

# Xylitol strain knockout model

# copy model to create mutant
mutant_xylitol = model.copy()

# create list of genes to KO
ko_list = pd.read_csv('data/df_xylitol_knockouts.csv', index_col=0).dropna(subset = ['id_bigg'])
ko_list.iML1515_idx = ko_list.iML1515_idx.astype(int)
#ko_list

for i in ko_list.iML1515_idx:
    mutant_xylitol.genes[i].knock_out()

xylitol_gene_ko = evaluate_knockout(model, mutant_xylitol, ko_list.iML1515_idx)
xylitol_gene_ko.to_csv('result/iML1515_xylitol_strain_ko_result.csv')
xylitol_gene_ko

### Evaluate Gene KOs
* Because one gene can be related to several reactions and vice versa, KO of one gene might not make the flux boundaries become 0. 
* First, we need to check if the knocked reaction is the one we intend to do
* If we decide to knockout all reactions related to the gene, then we can do the next line of codes:

# create list of genes to KO
ko_list = pd.read_csv('data/df_xylitol_knockouts.csv', index_col=0).dropna(subset = ['id_bigg'])
ko_list.iML1515_idx = ko_list.iML1515_idx.astype(int)
#ko_list

for count, i in enumerate(xylitol_gene_ko.ko_result):
    if not i:
        x = xylitol_gene_ko.iloc[count].reaction_id
        mutant_xylitol.reactions.get_by_id(x).knock_out()
xylitol_reaction_ko = evaluate_knockout(model, mutant_xylitol, ko_list.iML1515_idx)
xylitol_reaction_ko.to_csv('result/iML1515_xylitol_strain_reaction_ko_result.csv')
xylitol_reaction_ko

mutant_xylitol.reactions.CYTBDpp.bounds = (0.0, 1000.0)
mutant_xylitol.reactions.CYTBDpp.bounds

mutant_xylitol.reactions.XYLK.bounds

# Isobutyric acid strain knockout model

# copy model to create mutant
mutant_isobutyric = model.copy()

# create list of genes to KO
ko_list = pd.read_csv('data/df_isobutyric_knockouts.csv', index_col=0).dropna(subset = ['id_bigg'])
ko_list.iML1515_idx = ko_list.iML1515_idx.astype(int)
#ko_list

for i in ko_list.iML1515_idx:
    mutant_isobutyric.genes[i].knock_out()

isobutyric_gene_ko = evaluate_knockout(model, mutant_isobutyric, ko_list.iML1515_idx)
isobutyric_gene_ko.to_csv('result/iML1515_isobutyric_strain_ko_result.csv')   
isobutyric_gene_ko

### Evaluate Gene KOs
* Because one gene can be related to several reactions and vice versa, KO of one gene might not make the flux boundaries become 0. 
* First, we need to check if the knocked reaction is the one we intend to do
* If we decide to knockout all reactions related to the gene, then we can do the next line of codes:

# create list of genes to KO
ko_list = pd.read_csv('data/df_isobutyric_knockouts.csv', index_col=0).dropna(subset = ['id_bigg'])
ko_list.iML1515_idx = ko_list.iML1515_idx.astype(int)
#ko_list

for count, i in enumerate(isobutyric_gene_ko.ko_result):
    if not i:
        x = isobutyric_gene_ko.iloc[count].reaction_id
        mutant_isobutyric.reactions.get_by_id(x).knock_out()
isobutyric_reaction_ko = evaluate_knockout(model, mutant_isobutyric, ko_list.iML1515_idx)
isobutyric_reaction_ko.to_csv('result/iML1515_isobutyric_strain_reaction_ko_result.csv')
isobutyric_reaction_ko

# Add Reactions
In this section, we would like to add the heterologous reaction to the model. 
We need to crosscheck with the researchers whether this is the reaction intended.

## Load Model Yeast
We will copy some of the reactions from yeast model in BIGG

model_yeast = load_model('iMM904')
model_yeast.solver = 'gurobi'
model_yeast

## Define metabolites to be used in the reaction

# add metabolites
# only on isobutyric strain:
isobuta_c = Metabolite(
    'isobuta_c',
    formula='C4H8O2',
    name='2-methylpropanoic-acid',
    compartment='c')

isobuta_e = Metabolite(
    'isobuta_e',
    formula='C4H8O2',
    name='2-methylpropanoic-acid',
    compartment='e')

mppal_c = Metabolite(
    '2mppal_c',
    formula='C4H8O',
    name='2-methylpropanal',
    compartment='c')

# only on xylitol strain:
xylt_c = model_yeast.metabolites.xylt_c
xylt_e = model_yeast.metabolites.xylt_e

# general metabolites to add in reaction
xyl__D_c = model.metabolites.xyl__D_c
h_c = model.metabolites.h_c
co2_c = model.metabolites.co2_c
mob_c = model.metabolites.get_by_id('3mob_c')
h2o_c = model.metabolites.h2o_c
nad_c = model.metabolites.nad_c
nadh_c = model.metabolites.nadh_c
nadp_c = model.metabolites.nadp_c
nadph_c = model.metabolites.nadph_c

## Add Reaction to Isobutyric Mutant
### _Reaction 1: 3-Methyl-2-oxobutanoate => 2-methylpropanal_

# define reaction
reaction = Reaction('kivD_heterologous')
reaction.name = 'alpha-ketoisovalerate-decarboxylase'
reaction.subsystem = 'Alternate Carbon Metabolism'
reaction.lower_bound = 0.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    mob_c: -1.0,
    h_c: -1.0,
    co2_c: 1.0,
    mppal_c: 1.0
})

reaction.reaction  

reaction.gene_reaction_rule = '( kivD )'
reaction.genes

mutant_isobutyric.add_reactions([reaction])

### _Reaction 2: 2-methylpropanal =>  2-methylpropanoic-acid_

# define reaction
reaction = Reaction('ALDD19xr_mppal')
reaction.name = 'Aldehyde dehydrogenase (2-methylpropanal, NAD)'
reaction.subsystem = 'Alternate Carbon Metabolism'
reaction.lower_bound = -1000.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    h2o_c: -1.0,
    nad_c: -1.0, 
    mppal_c: -1.0,
    h_c: 2.0,
    nadh_c: 1.0,
    isobuta_c: 1.0
})

reaction.reaction  

reaction.gene_reaction_rule = '( b1385 )'
reaction.genes

mutant_isobutyric.add_reactions([reaction])

### _Reaction 3: 2-methylpropanoic-acid transport_

# define reaction
reaction = Reaction('isobutat')
reaction.name = '2-methylpropanoic-acid transport'
reaction.subsystem = 'Transport, Extracellular' #TRANSPORT
reaction.lower_bound = -1000.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    isobuta_c: -1.0,
    isobuta_e: 1.0
})

reaction.reaction  

reaction.gene_reaction_rule = '( isobuta_transporter )'
reaction.genes

mutant_isobutyric.add_reactions([reaction])

### _Reaction 4: 2-methylpropanoic-acid => exchange_

# define reaction
reaction = Reaction('EX_isobuta_e')
reaction.name = '2-methylpropanoic-acid exchange'
reaction.subsystem = 'Extracellular exchange'
reaction.lower_bound = 0.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    isobuta_e: -1.0
})

reaction.reaction  

mutant_isobutyric.add_reactions([reaction])

### Add metabolites from other mutant to enable comparison of exchanges

mutant_isobutyric.add_metabolites(xylt_e)

reaction_1 = mutant_isobutyric.reactions.get_by_id('BIOMASS_Ec_iML1515_core_75p37M')
#reaction_1 = mutant.reactions.get_by_id('ATPM')
reaction_2 = mutant_isobutyric.reactions.get_by_id('EX_isobuta_e')

new_objective = {reaction_1: 0.93, reaction_2:0.07}

mutant_isobutyric.objective = new_objective

mutant_isobutyric.objective.expression

### Write Model

mutant_isobutyric.id = mutant_isobutyric.id+"_iso_obj"
cobra.io.write_sbml_model(mutant_isobutyric, "result/iML1515_isobutyric_strain.xml")
cobra.io.save_json_model(mutant_isobutyric, "result/iML1515_isobutyric_strain.json")

## Xylitol Mutant
### _Reaction 1: NADPH-dependent D-xylose reductase_

# THIS REACTION CAN BE FOUND FROM THE YEAST MODEL IN BIGG
# https://www.uniprot.org/uniprot/Q8X195 from Candida boidinii (Yeast)
# model reaction modified from http://bigg.ucsd.edu/models/iMM904/reactions/XYLR
# need to check if its the same with uniprot http://bigg.ucsd.edu/models/iMM904/genes/YHR104W
reaction = Reaction('XYLR')
reaction.name = 'NADPH-dependent D-xylose reductase'
reaction.subsystem = 'S_Xylose_Metabolism'
reaction.lower_bound = 0.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    h_c: -1.0,
    nadp_c: 1.0,
    nadph_c: -1.0,
    xyl__D_c: -1.0,
    xylt_c: 1.0,
})

reaction.reaction  

reaction.gene_reaction_rule = '( XYL1 )'
reaction.genes

reaction = model_yeast.reactions.XYLR

mutant_xylitol.add_reactions([reaction])
mutant_xylitol.reactions.XYLR.bounds = (0.0, 1000.0)

### _Reaction 2: Xylitol transport via passive diffusion_

reaction = model_yeast.reactions.XYLTt
reaction.reaction

mutant_xylitol.add_reactions([reaction])
mutant_xylitol.reactions.XYLTt.bounds = (-1000.0, 1000.0)

### _Reaction 3: Xylitol Exchange_

# define reaction
reaction = Reaction('EX_xylt_e')
reaction.name = 'Xylitol Exchange'
reaction.subsystem = 'Extracellular exchange'
reaction.lower_bound = 0.  # This is the default
reaction.upper_bound = 1000.  # This is the default

reaction.add_metabolites({
    xylt_e: -1.0
})

reaction.reaction 

mutant_xylitol.add_reactions([reaction])

### Add metabolites from other mutant to enable comparison of exchanges

mutant_xylitol.add_metabolites(isobuta_e)

# Redefine Objective?

reaction_1 = mutant_xylitol.reactions.get_by_id('BIOMASS_Ec_iML1515_core_75p37M')
#reaction_1 = mutant.reactions.get_by_id('ATPM')
reaction_2 = mutant_xylitol.reactions.get_by_id('EX_xylt_e')
reaction_3 = mutant_xylitol.reactions.get_by_id('EX_ac_e')

new_objective = {reaction_1: 0.700, reaction_2:0.290, reaction_3:0.010}
mutant_xylitol.objective = new_objective

### Write Model

mutant_xylitol.id = mutant_xylitol.id+"_xy_obj"
cobra.io.write_sbml_model(mutant_xylitol, "result/iML1515_xylitol_strain.xml")
cobra.io.save_json_model(mutant_xylitol, "result/iML1515_xylitol_strain.json")

mutant_xylitol.reactions.get_by_id("CYTBDpp").knock_out()
cobra.io.write_sbml_model(mutant_xylitol, "result/iML1515_xylitol_strain_crispri.xml")
cobra.io.save_json_model(mutant_xylitol, "result/iML1515_xylitol_strain_crispri.json")

mutant_xylitol.objective.expression

# [CLICK FOR NEXT STEP](01-0_simulation_xylitol_strain.ipynb)
