==> 01-1_simulation_isobutyric_strain.ipynb <==
#### _Matin's Lab Rotation 3 - BCFO_ | [NEXT](03_community_dFBA_growth_objective) | [PREVIOUS](01-0_simulation_xylitol_strain.ipynb)
# 03. Simulation of the Xylitol Producing Strain
In this section, we would like to see if the model can:
* Uptake Carbon: Glucose and Xylose
* Grow: Biomass & ATP
* Produce: Acetate & Xylitol

## Load Library

from cameo import load_model
from cameo import fba
from escher import Builder
import pandas as pd
import cobra

def simulate_model(experiment_id, model, objective, target, metabolite):
    model.objective = objective
    solution = fba(model)
    f_data = solution.fluxes
    f_data = f_data.rename(experiment_id)
    solution_result = []
    for i in target:
        flux = solution.fluxes.get(i)
        solution_result.append(flux)
    data = [experiment_id] + solution_result
    #print(data)
    df_target = pd.DataFrame(columns=[["Model"]+target], data=[data])
    print("experiment:", experiment_id, "| evaluate compound:", model.metabolites.get_by_id(metabolite).summary())
    return df_target, f_data

def run_series_simulation(experiment, model, mutant, target, metabolite):
    solution_fluxes = pd.DataFrame()
    df = pd.DataFrame()
    for i in experiment.index:
        experiment_id = experiment.loc[i, "experiment_id"]
        if experiment.loc[i, "model_id"] != "model":
            m = mutant
        else:
            m = model
        objective = experiment.loc[i, "objective"]
        data, sol = simulate_model(experiment_id, m, objective, target, metabolite)
        solution_fluxes = solution_fluxes.append(sol)
        df = df.append(data)
        print("****************************************", "\n")
    return df, solution_fluxes

def visualize_flux(model, objective, map_data='map/iJO1366.Central metabolism_rotation3.json'):
    model.objective = objective
    flux = fba(model).fluxes
    builder = Builder(
        map_json = map_data,
        reaction_data = flux
    )
    return builder

## Load Model

# WT for comparison
model1 = load_model('iML1515')
model1.solver = 'gurobi'
model1

# load mutant
mutant = load_model('result/iML1515_isobutyric_strain.xml')
mutant.solver = 'gurobi'
mutant

## Define Media: Glucose & Xylose

# define media composition
media_components = [["EX_xyl__D_e", 75],
                    ["EX_glc__D_e", 75],
                    ["EX_ac_e", 75],
                    ["EX_o2_e", 1000]
                   ]

# change the media in the model
medium = model1.medium
for i in media_components:
    medium[i[0]] = i[1]
    model1.medium = medium
#print(model.medium)

# change the media in the mutant
medium = mutant.medium
for i in media_components:
    medium[i[0]] = i[1]
    mutant.medium = medium
#print(mutant.medium)

# WT Grow on Glu & Xyl

## WT Biomass Objective

visualize_flux(model1, "BIOMASS_Ec_iML1515_core_75p37M")

## WT under Anaerobic Condition

media_components = [["EX_o2_e", 0]]
medium = model1.medium
for i in media_components:
    medium[i[0]] = i[1]
    model1.medium = medium
visualize_flux(model1, "BIOMASS_Ec_iML1515_core_75p37M")

# Mutant Growth on Glu & Xyl
## Mt Growth Objective

visualize_flux(mutant, "BIOMASS_Ec_iML1515_core_75p37M")

## Mt Xylitol Objective

reaction_1 = mutant.reactions.get_by_id('BIOMASS_Ec_iML1515_core_75p37M')
#reaction_1 = mutant.reactions.get_by_id('ATPM')
reaction_2 = mutant.reactions.get_by_id('EX_isobuta_e')

new_objective = {reaction_1: 0.93, reaction_2:0.07}
visualize_flux(mutant, new_objective)

mutant.summary()

mutant.id = mutant.id+"_xy_strain_crispr_off"
cobra.io.write_sbml_model(mutant, "result/model_for_consortia/iML1515_xylitol_strain_crispr_off.xml")
cobra.io.save_json_model(mutant, "result/model_for_consortia/iML1515_xylitol_strain_crispr_off.json")
mutant.metabolites.ac_c.summary()

visualize_flux(crispri(mutant, "on"), new_objective)

mutant.summary()

mutant.id = mutant.id+"_xy_strain_crispr_on"
cobra.io.write_sbml_model(mutant, "result/model_for_consortia/iML1515_xylitol_strain_crispr_on.xml")
cobra.io.save_json_model(mutant, "result/model_for_consortia/iML1515_xylitol_strain_crispr_on.json")
mutant.metabolites.ac_c.summary()
