==> 02_consortia_model.ipynb <==
#### _Matin's Lab Rotation 3 - BFCO_ | [NEXT](03_community_dFBA_growth_objective.ipynb) | [PREVIOUS](01-1_simulation_isobutyric_strain.ipynb)
# 0X. Test acetate exchange

from cameo import load_model
from cameo import fba
from escher import Builder
import pandas as pd
import cobra

model = load_model('iML1515')
model.solver = 'gurobi'
model

mutant1 = load_model('result/iML1515_xylitol_strain.xml')
mutant1.solver = 'gurobi'
mutant1

mutant2 = load_model('result/iML1515_isobutyric_strain.xml')
mutant2.solver = 'gurobi'
mutant2

media_components = [["EX_xyl__D_e", 5.0],
                    ["EX_glc__D_e", 5.0],
                   ]

medium = mutant1.medium
for i in media_components:
    medium[i[0]] = i[1]
    mutant1.medium = medium
#print(model.medium)

mutant2.medium = mutant1.medium
model.medium = mutant2.medium

reaction_1 = mutant1.reactions.get_by_id('BIOMASS_Ec_iML1515_core_75p37M')
reaction_2 = mutant1.reactions.get_by_id('EX_xylt_e')
reaction_3 = mutant1.reactions.get_by_id('EX_ac_e')
mutant1_new_objective = {reaction_1: 0.8, reaction_2:0.18, reaction_3:0.02}
mutant1.objective = mutant1_new_objective
mutant1.reactions.XYLK.knock_out()
mutant1.genes.b0733.knock_out()
#cobra.io.write_sbml_model(mutant1, "result/model_for_consortia/iML1515_mutant1.xml")

#reaction_1 = mutant2.reactions.get_by_id('BIOMASS_Ec_iML1515_core_75p37M')
reaction_4 = mutant2.reactions.get_by_id('EX_isobuta_e')
mutant2_new_objective = {reaction_1: 0.9, reaction_4:0.1}
mutant2.objective = mutant2_new_objective
#cobra.io.write_sbml_model(mutant2, "result/model_for_consortia/iML1515_mutant2.xml")

solution0 = model.optimize()

solution1 = mutant1.optimize()

ac_e = mutant1.reactions.EX_ac_e.summary().to_frame().values[0][0]
ac_e
media_components = [["EX_ac_e", ac_e]]
medium = mutant2.medium
for i in media_components:
    medium[i[0]] = i[1]
    mutant2.medium = medium

solution2 = mutant2.optimize()

builder = Builder(
    map_json='map/iJO1366.Central metabolism_rotation3.json',
    reaction_data=solution0.fluxes,
    #reaction_data=solution.fluxes,
)
builder.save_html('result/flux_map/model.html')
builder

builder = Builder(
    map_json='map/iJO1366.Central metabolism_rotation3.json',
    reaction_data=solution1.fluxes,
    #reaction_data=solution.fluxes,
)
builder.save_html('result/flux_map/mutant1.html')
builder

#print(mutant1.reactions.BIOMASS_Ec_iML1515_core_75p37M.summary())
mutant1.summary()

builder = Builder(
    map_json='map/iJO1366.Central metabolism_rotation3.json',
    reaction_data=solution2.fluxes,
    #reaction_data=solution.fluxes,
)
builder.save_html('result/flux_map/mutant2.html')
builder

mutant2.summary()

cobra.manipulation.delete_model_genes(mutant2, ['b0733'])
solution3 = mutant2.optimize()

builder = Builder(
    map_json='map/iJO1366.Central metabolism_rotation3.json',
    reaction_data=solution3.fluxes,
    #reaction_data=solution.fluxes,
)
builder

data0 =[solution0.fluxes.BIOMASS_Ec_iML1515_core_75p37M,
        solution0.fluxes.EX_co2_e, 
        solution0.fluxes.EX_ac_e,
        #solution1.fluxes.EX_xylt_e
       ]
print(data0)

data1 =[solution1.fluxes.BIOMASS_Ec_iML1515_core_75p37M,
        solution1.fluxes.EX_co2_e, 
        solution1.fluxes.EX_ac_e,
        solution1.fluxes.EX_xylt_e
       ]
print(data1)

data2 =[solution2.fluxes.BIOMASS_Ec_iML1515_core_75p37M,      
        solution2.fluxes.EX_co2_e,
        solution2.fluxes.EX_ac_e,
        solution2.fluxes.EX_isobuta_e,
]
print(data2)
